requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==9.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
